added a utils file
I tried adding handle keys and render also but it was not working
