+++ frontend/src/components/sudoku/Board9x9.jsx
@@ -1,30 +1,16 @@
-import { useState, useEffect, useCallback } from "react";
+import { useEffect, useCallback } from "react";
 import Cell from "./Cell";
 import Keypad from "./Keypad";
 import { getNineBoard } from "../../api/getBoard";
-
+import { useSudokuGrid, useSelectedCell } from "./BoardUtils";
 function Board9x9() {
-  const [sudokuGrid, setSudokuGrid] = useState(Array.from({ length: 9 }, () => Array(9).fill("")));
-  const [selectedCell, setSelectedCell] = useState({ row: 0, col: 0 });
+  const { sudokuGrid, handleCellChange } = useSudokuGrid(9);
+  const { selectedCell, setSelectedCell, handleCellClick } = useSelectedCell();
 
   useEffect(() => {
     console.log("getNineBoard");
     getNineBoard().then((data) => console.log(data));
   }, []);
-
-  const handleCellChange = useCallback(
-    (row, col, value) => {
-      const newGrid = [...sudokuGrid];
-      newGrid[row][col] = value;
-      setSudokuGrid(newGrid);
-    },
-    [sudokuGrid],
-  );
-
-  const handleCellClick = (row, col) => {
-    console.log(`Selected cell: (${row + 1}, ${col + 1})`);
-    setSelectedCell({ row, col });
-  };
 
   const handleKeypadClick = (value) => {
     if (selectedCell.row !== null && selectedCell.col !== null) {
